# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type AnalyzeNutritionRequestModel = typeof AnalyzeNutritionRequestModel[keyof typeof AnalyzeNutritionRequestModel];


export const AnalyzeNutritionRequestModel = {
  auto: 'auto',
  openai: 'openai',
  gemini: 'gemini',
  claude: 'claude',
} as const;

export interface AnalyzeNutritionRequest {
  imageBase64: string;
  model?: AnalyzeNutritionRequestModel;
}

export interface BilingualText {
  en: string;
  ar: string;
}

export interface NutritionMacros {
  protein: number;
  fat: number;
  saturatedFat: number;
  carbohydrates: number;
  sugar: number;
  fiber: number;
}

export interface NutritionMinerals {
  sodium: number;
  calcium: number;
  iron: number;
}

export interface NutritionWarnings {
  diabetes: boolean;
  bloodPressure: boolean;
  allergies: string[];
  kidney: boolean;
  lowCarb: boolean;
  glutenSensitivity: boolean;
  lactoseIntolerance: boolean;
  highFat: boolean;
  highSugar: boolean;
}

export interface NutritionData {
  foodName: BilingualText;
  description: BilingualText;
  servingSize: string;
  calories: number;
  macros: NutritionMacros;
  minerals: NutritionMinerals;
  vitamins: string[];
  healthScore: number;
  imageQuality: string;
  isEstimated: boolean;
  healthBenefits: BilingualText;
  healthCautions: BilingualText;
  warnings: NutritionWarnings;
}

export interface StoredAnalysis {
  id: string;
  nutritionData: NutritionData;
  modelUsed: string;
  imageThumb?: string | null;
  createdAt: string;
}