# lib/api-zod/src/generated/types/analyzeNutritionRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeNutritionRequestModel } from './analyzeNutritionRequestModel';

export interface AnalyzeNutritionRequest {
  imageBase64: string;
  model?: AnalyzeNutritionRequestModel;
}