# lib/api-zod/src/generated/types/analyzeNutritionRequestModel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type AnalyzeNutritionRequestModel = typeof AnalyzeNutritionRequestModel[keyof typeof AnalyzeNutritionRequestModel];


export const AnalyzeNutritionRequestModel = {
  auto: 'auto',
  openai: 'openai',
  gemini: 'gemini',
  claude: 'claude',
} as const;