# lib/api-zod/src/generated/types/nutritionData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BilingualText } from './bilingualText';
import type { NutritionMacros } from './nutritionMacros';
import type { NutritionMinerals } from './nutritionMinerals';
import type { NutritionWarnings } from './nutritionWarnings';

export interface NutritionData {
  foodName: BilingualText;
  description: BilingualText;
  servingSize: string;
  calories: number;
  macros: NutritionMacros;
  minerals: NutritionMinerals;
  vitamins: string[];
  healthScore: number;
  imageQuality: string;
  isEstimated: boolean;
  healthBenefits: BilingualText;
  healthCautions: BilingualText;
  warnings: NutritionWarnings;
}