# lib/api-zod/src/generated/types/nutritionWarnings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NutritionWarnings {
  diabetes: boolean;
  bloodPressure: boolean;
  allergies: string[];
  kidney: boolean;
  lowCarb: boolean;
  glutenSensitivity: boolean;
  lactoseIntolerance: boolean;
  highFat: boolean;
  highSugar: boolean;
}