# lib/api-zod/src/generated/types/storedAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NutritionData } from './nutritionData';

export interface StoredAnalysis {
  id: string;
  nutritionData: NutritionData;
  modelUsed: string;
  imageThumb?: string | null;
  createdAt: Date;
}